[FIR] Rewrite `findStaticallyKnownSubtype()`
In `isErasedAllowForTypeWithIrrelevantMixin`, at some point we come to
`isSubtypeOfForSingleClassifierType` where we collect supertypes of
`DerivedSuper` and take their arguments (`TypeVariable(S)` and `D`) and
intersect them. This results in `(TypeVariable(S) & Any) & D`, for which
inference later deduces:

1. `(TypeVariable(S) & Any) & D <: B`
  - in `simplifyConstraintForPossibleIntersectionSubType` we split
    "our type variables" form other types, thus considering
    `TypeVariable(S) & Any` and `D` separately. Because `D <: Any`,
    we relax the constraints for our type variables, allowing them
    to be nullable
2. `TypeVariable(S) & Any <: B?`
  - adding a constraint to a DNN means adding its nullable version to
    the original variable, so
3. `TypeVariable(S) <: B?`

All this leads to deriving `DerivedSuper<out B?>` as the statically
known type. Informally, we are allowed to consider it
`BaseSuper<out B?>` because `BaseSuper<D>` already allows us to conclude
`BaseSuper<out B>` from it if we need it.

^KT-7972 Fixed