FIR: Support using implicit context values for FunctionN.invoke
Namely, for the case when there's a property `f` of a function type
in a form of `context(C1, .., CN) (V1, .. , VM) -> ReturnType`, but
at call-sites, there's only M explicit arguments `f(e1, .., em)`,
and at there are enough implicit context values to supply for C1, .., CN.

For expected behavior, see §1.7 at
https://github.com/Kotlin/KEEP/blob/aff2891bf9d997fdddcc3567c95a2553f0503ac9/proposals/context-parameters.md

The basic implementation idea is the following:
- At MapArguments, when the number of explicit value arguments
is smaller or equal than
`allValueParameterCount - contextParameterCount`,
add stub value arguments with expected context value types

- At CheckContextArguments, run a regular context resolution algorithm
and replace stub arguments with found values if it's successful.

But there are, of course, some tricky details described in the code.

^KT-73331 Fixed
^KT-51383 Fixed
^KT-54068 Fixed
^KT-73805 Related