[Analysis API] introduce `KotlinResolutionHelperService`
This service provides a bridge between the Analysis API, Kotlin IntelliJ
plugin and Java IntelliJ plugin. In particular, it is indented to adjust
the behavior of the Java resolver with `PsiFileEx.BATCH_REFERENCE_PROCESSING`
flag to avoid an aggressive cache strategy to comply with the Kotlin
compiler phase contracts when the Kotlin resolver is on the stack.

^KT-73649