Tests: disable linkage mode test check when on FIR and with signatures
Signatures support is dropped from the FIR pipeline. Therefore, the test
makes no sense in this mode now. In fact, it is not doing any checks
in the current master because the pipeline it uses
(GenerationUtils,compileFilesUsingFrontendIR) ignores installed
generator extensions anyway.
But since I put the extensions back there to restore the pipeline for
some old scripting tests, after moving scripting lowering to the plugin,
the linkage test started to fail.