Tests: drop obsolete Kapt4 test that starts failing after changes
... in the test pipeline.
After pipeline refactorings this test uses the obsolete compilation
pipeline, different for the CLI one, so it looks like it lost its
purpose. On the other hand, since I put the generation extensions back
in the test pipeline to restore its functionality for some old
scripting tests, after moving scripting lowering to the plugin,
this Kapt test started to fail.